Skapade en buss katalog med linje objekt. Funderar på att Bryta ut Linje delen till ett eget system så att vi kan använda den till X objekt t.ex taxi, båt
