+++ buss/BussLinesCollection.py
@@ -1,0 +1,17 @@
+class BussLinesCollection:
+    def __init__(self):
+        with open("busslinje.txt", "r") as f:
+            obs = f.readlines()
+
+            self.linje = []
+            for d in obs:
+                self.linje.append(d)
+
+    def get_bussline_by_id(self, id):
+            return self.linje[id]
+
+        # läs in en linje från filen
+        # skapa alla hållplatser för denna linje
+        # för varje hållplats anropa add_stop för denna linje
+        # När linjen är färdigskapad anropa add_line och skicka med denna linje
+        # stäng filen när klar
