+++ class5.py
@@ -20,195 +20,6 @@ class BussDriverCollection:
 #########################################################################################
 
 
-class BussLinesCollection:
-    def __init__(self):
-        with open("busslinje.txt", "r") as f:
-            obs = f.readlines()
-
-            self.linje = []
-            for d in obs:
-                self.linje.append(d)
-
-    def get_bussline_by_id(self, id):
-            return self.linje[id]
-
-        # läs in en linje från filen
-        # skapa alla hållplatser för denna linje
-        # för varje hållplats anropa add_stop för denna linje
-        # När linjen är färdigskapad anropa add_line och skicka med denna linje
-        # stäng filen när klar
-
-
-class Buss:
-    def __init__(self):
-        pass
-
-    def samla_info(self, driver1, report, status, linje, avg):
-        print(driver1, report, status, linje, avg)
-        c=input("d")
-
-    def Traffic_addstuff(self, driver1, report, status):
-        print("Välj vilken linje du sitter på?")
-        print("****************************************************")
-        print(f"""1.{BussLinesCollection().get_bussline_by_id(0)}""")
-        print(f"""2.{BussLinesCollection().get_bussline_by_id(1)}""")
-        print(f"""3.{BussLinesCollection().get_bussline_by_id(2)}""")
-        linje = input("Choose one making you better feeling")
-
-        if linje == "1":
-            valdlinje = BussLinesCollection().get_bussline_by_id(0)
-            print(f"""Välj avgång för busslinjen:""")
-            print("*********************************")
-            print(f"""1.{Timetable().get_timetable1_spec(0)}""")
-            print(f"""2.{Timetable().get_timetable1_spec(1)}""")
-            print(f"""3.{Timetable().get_timetable1_spec(2)}""")
-            print(f"""4.{Timetable().get_timetable1_spec(3)}""")
-            print(f"""5.{Timetable().get_timetable1_spec(4)}""")
-            print(f"""6.{Timetable().get_timetable1_spec(5)}""")
-            timetable1 = input("Choose specific departure:")
-            if timetable1 == "1":
-                valdavg = Timetable().get_timetable1_spec(0)
-                print(f"""{Timetable().get_timetable1_spec(0)}""")
-                Buss().samla_info(driver1, report, status, valdlinje, valdavg)
-            elif timetable1 == "2":
-                valdavg = Timetable().get_timetable1_spec(1)
-                print(f"""{Timetable().get_timetable1_spec(1)}""")
-                Buss().samla_info(driver1, report, status, valdlinje, valdavg)
-            elif timetable1 == "3":
-                valdavg = Timetable().get_timetable1_spec(2)
-                print(f"""{Timetable().get_timetable1_spec(2)}""")
-                Buss().samla_info(driver1, report, status, valdlinje, valdavg)
-            elif timetable1 == "4":
-                valdavg = Timetable().get_timetable1_spec(3)
-                print(f"""{Timetable().get_timetable1_spec(3)}""")
-                Buss().samla_info(driver1, report, status, valdlinje, valdavg)
-            elif timetable1 == "5":
-                valdavg = Timetable().get_timetable1_spec(4)
-                print(f"""{Timetable().get_timetable1_spec(4)}""")
-                Buss().samla_info(driver1, report, status, valdlinje, valdavg)
-            elif timetable1 == "6":
-                valdavg = Timetable().get_timetable1_spec(5)
-                print(f"""{Timetable().get_timetable1_spec(5)}""")
-                Buss().samla_info(driver1, report, status, valdlinje, valdavg)
-            else:
-                print("fuck u")
-
-        elif linje == "2":
-            valdlinje = BussLinesCollection().get_bussline_by_id(1)
-            print(f"""Välj avgång för busslinjen:""")
-            print("*********************************")
-            print(f"""1.{Timetable().get_timetable2_spec(0)}""")
-            print(f"""2.{Timetable().get_timetable2_spec(1)}""")
-            print(f"""3.{Timetable().get_timetable2_spec(2)}""")
-            print(f"""4.{Timetable().get_timetable2_spec(3)}""")
-            print(f"""5.{Timetable().get_timetable2_spec(4)}""")
-            print(f"""6.{Timetable().get_timetable2_spec(5)}""")
-            timetable2 = input("Choose specific departure:")
-            if timetable2 == "1":
-                valdavg = Timetable().get_timetable2_spec(0)
-                print(f"""{Timetable().get_timetable2_spec(0)}""")
-                Buss().samla_info(driver1, report, status, valdlinje, valdavg)
-            elif timetable2 == "2":
-                valdavg = Timetable().get_timetable2_spec(1)
-                print(f"""{Timetable().get_timetable2_spec(1)}""")
-                Buss().samla_info(driver1, report, status, valdlinje, valdavg)
-            elif timetable2 == "3":
-                valdavg = Timetable().get_timetable2_spec(2)
-                print(f"""{Timetable().get_timetable2_spec(2)}""")
-                Buss().samla_info(driver1, report, status, valdlinje, valdavg)
-            elif timetable2 == "4":
-                valdavg = Timetable().get_timetable2_spec(3)
-                print(f"""{Timetable().get_timetable2_spec(3)}""")
-                Buss().samla_info(driver1, report, status, valdlinje, valdavg)
-            elif timetable2 == "5":
-                valdavg = Timetable().get_timetable2_spec(4)
-                print(f"""{Timetable().get_timetable2_spec(4)}""")
-                Buss().samla_info(driver1, report, status, valdlinje, valdavg)
-            elif timetable2 == "6":
-                valdavg = Timetable().get_timetable2_spec(5)
-                print(f"""{Timetable().get_timetable2_spec(5)}""")
-                Buss().samla_info(driver1, report, status, valdlinje, valdavg)
-            else:
-                print("fuck u")
-
-        elif linje == "3":
-            valdlinje = BussLinesCollection().get_bussline_by_id(2)
-            print(f"""Välj avgång för busslinjen:""")
-            print("*********************************")
-            print(f"""1.{Timetable().get_timetable3_spec(0)}""")
-            print(f"""2.{Timetable().get_timetable3_spec(1)}""")
-            print(f"""3.{Timetable().get_timetable3_spec(2)}""")
-            print(f"""4.{Timetable().get_timetable3_spec(3)}""")
-            print(f"""5.{Timetable().get_timetable3_spec(4)}""")
-            print(f"""6.{Timetable().get_timetable3_spec(5)}""")
-            timetable3 = input("Choose specific departure:")
-            if timetable3 == "1":
-                valdavg = Timetable().get_timetable3_spec(0)
-                print(f"""{Timetable().get_timetable3_spec(0)}""")
-                Buss().samla_info(driver1, report, status, valdlinje, valdavg)
-            elif timetable3 == "2":
-                valdavg = Timetable().get_timetable3_spec(1)
-                print(f"""{Timetable().get_timetable3_spec(1)}""")
-                Buss().samla_info(driver1, report, status, valdlinje, valdavg)
-            elif timetable3 == "3":
-                valdavg = Timetable().get_timetable3_spec(2)
-                print(f"""{Timetable().get_timetable3_spec(2)}""")
-                Buss().samla_info(driver1, report, status, valdlinje, valdavg)
-            elif timetable3 == "4":
-                valdavg = Timetable().get_timetable3_spec(3)
-                print(f"""{Timetable().get_timetable3_spec(3)}""")
-                Buss().samla_info(driver1, report, status, valdlinje, valdavg)
-            elif timetable3 == "5":
-                valdavg = Timetable().get_timetable3_spec(4)
-                print(f"""{Timetable().get_timetable3_spec(4)}""")
-                Buss().samla_info(driver1, report, status, valdlinje, valdavg)
-            elif timetable3 == "6":
-                valdavg = Timetable().get_timetable3_spec(5)
-                print(f"""{Timetable().get_timetable3_spec(5)}""")
-                Buss().samla_info(driver1, report, status, valdlinje, valdavg)
-            else:
-                print("fuck u")
-        else:
-            print("Choose another option:")
-        # Meny för själva linjerna till bussarna
-
-
-class Linjemenu:
-    def __init__(self):
-        self.choices = {
-            "1": BussLinesCollection().get_bussline_by_id(0),
-            "2": BussLinesCollection().get_bussline_by_id(1),
-            "3": BussLinesCollection().get_bussline_by_id(2),
-        }
-
-    def display_linjemenu(self):
-        print(f"""
-Busslinjer
-** ** ** ** ** ** ** ** ** ** ** **
-1.{BussLinesCollection().get_bussline_by_id(0)}
-2.{BussLinesCollection().get_bussline_by_id(1)}
-3.{BussLinesCollection().get_bussline_by_id(2)}
-        """)
-
-    def run(self):
-        while True:
-            self.display_linjemenu()
-            choice = input("Välj linje för se tidtabell: ")
-            action = choice
-            if action == "1":
-                print("""\nGöteborg Centralstationen - Uddevalla Kampenhof\n** ** ** ** ** ** ** ** ** ** ** **""")
-                Timetable().get_timetable1()
-            elif action == "2":
-                print("""\nPartille Centrum - Nordstan\n** ** ** ** ** ** ** ** ** ** ** **""")
-                Timetable().get_timetable2()
-
-            elif action == "3":
-                print("""\nKungsbacka Station - Göteborg Centralstation\n** ** ** ** ** ** ** ** ** ** ** **""")
-                Timetable().get_timetable3()
-
-            else:
-                print("is not an option".format(choice))
-
 
 class TrafficMenu:
     def __init__(self):
@@ -305,7 +116,7 @@ class Menu:
             print("Wrong password")
 
     def consumer(self):
-        Linjemenu().run()
+        BussLineMenu().run()
 
 class Get_time:
     def __init__(self, avg, ank):
